0 - client - Update mutations and queries due to new version, reorganize
