+++ client/src/mutations/connectionTags.mutations.ts
@@ -1,0 +1,21 @@
+import {
+    ConnectionTagApi,
+    UpdateConnectionTagsRequest,
+} from '@/middleware/helios/connection';
+import {useMutation} from '@tanstack/react-query';
+
+type UpdateConnectionTagsMutationProps = {
+    onSuccess?: (result: void, variables: UpdateConnectionTagsRequest) => void;
+    onError?: (error: object, variables: UpdateConnectionTagsRequest) => void;
+};
+
+export const useUpdateConnectionTagsMutation = (
+    mutationProps?: UpdateConnectionTagsMutationProps
+) =>
+    useMutation<void, Error, UpdateConnectionTagsRequest>({
+        mutationFn: (request: UpdateConnectionTagsRequest) => {
+            return new ConnectionTagApi().updateConnectionTags(request);
+        },
+        onError: mutationProps?.onError,
+        onSuccess: mutationProps?.onSuccess,
+    });
