+++ client/src/mutations/connections.mutations.ts
@@ -1,9 +1,4 @@
-import {
-    ConnectionApi,
-    ConnectionModel,
-    ConnectionTagApi,
-    UpdateConnectionTagsRequest,
-} from '@/middleware/helios/connection';
+import {ConnectionApi, ConnectionModel} from '@/middleware/helios/connection';
 import {useMutation} from '@tanstack/react-query';
 
 type CreateConnectionMutationProps = {
@@ -56,22 +51,6 @@ export const useUpdateConnectionMutation = (
                 connectionModel: connection,
                 id: connection.id!,
             });
-        },
-        onError: mutationProps?.onError,
-        onSuccess: mutationProps?.onSuccess,
-    });
-
-type UpdateConnectionTagsMutationProps = {
-    onSuccess?: (result: void, variables: UpdateConnectionTagsRequest) => void;
-    onError?: (error: object, variables: UpdateConnectionTagsRequest) => void;
-};
-
-export const useUpdateConnectionTagsMutation = (
-    mutationProps?: UpdateConnectionTagsMutationProps
-) =>
-    useMutation<void, Error, UpdateConnectionTagsRequest>({
-        mutationFn: (request: UpdateConnectionTagsRequest) => {
-            return new ConnectionTagApi().updateConnectionTags(request);
         },
         onError: mutationProps?.onError,
         onSuccess: mutationProps?.onSuccess,
