+++ client/src/mutations/projectTags.mutations.ts
@@ -1,0 +1,21 @@
+import {
+    ProjectTagApi,
+    UpdateProjectTagsRequest,
+} from '@/middleware/helios/configuration';
+import {useMutation} from '@tanstack/react-query';
+
+type UpdateProjectTagsMutationProps = {
+    onSuccess?: (result: void, variables: UpdateProjectTagsRequest) => void;
+    onError?: (error: object, variables: UpdateProjectTagsRequest) => void;
+};
+
+export const useUpdateProjectTagsMutation = (
+    mutationProps?: UpdateProjectTagsMutationProps
+) =>
+    useMutation({
+        mutationFn: (request: UpdateProjectTagsRequest) => {
+            return new ProjectTagApi().updateProjectTags(request);
+        },
+        onError: mutationProps?.onError,
+        onSuccess: mutationProps?.onSuccess,
+    });
