+++ client/src/mutations/projectWorkflows.mutations.ts
@@ -1,0 +1,26 @@
+import {
+    CreateProjectWorkflowRequest,
+    ProjectApi,
+    WorkflowModel,
+} from '@/middleware/helios/configuration';
+import {useMutation} from '@tanstack/react-query';
+
+type CreateProjectWorkflowMutationProps = {
+    onSuccess?: (
+        result: WorkflowModel,
+        variables: CreateProjectWorkflowRequest
+    ) => void;
+    onError?: (error: object, variables: CreateProjectWorkflowRequest) => void;
+};
+
+export const useCreateProjectWorkflowMutation = (
+    mutationProps?: CreateProjectWorkflowMutationProps
+) => {
+    return useMutation({
+        mutationFn: (request: CreateProjectWorkflowRequest) => {
+            return new ProjectApi().createProjectWorkflow(request);
+        },
+        onError: mutationProps?.onError,
+        onSuccess: mutationProps?.onSuccess,
+    });
+};
