+++ client/src/mutations/projects.mutations.ts
@@ -1,16 +1,8 @@
 import {useMutation} from '@tanstack/react-query';
 import {
-    CreateProjectWorkflowRequest,
-    DeleteProjectWorkflowRequest,
-    DuplicateWorkflowRequest,
     ProjectApi,
     ProjectModel,
-    ProjectTagApi,
     PublishProjectRequest,
-    UpdateProjectTagsRequest,
-    UpdateWorkflowRequest,
-    WorkflowApi,
-    WorkflowModel,
 } from 'middleware/helios/configuration';
 
 type CreateProjectMutationProps = {
@@ -31,26 +23,6 @@ export const useCreateProjectMutation = (
         onSuccess: mutationProps?.onSuccess,
     });
 
-type CreateProjectWorkflowRequestMutationProps = {
-    onSuccess?: (
-        result: WorkflowModel,
-        variables: CreateProjectWorkflowRequest
-    ) => void;
-    onError?: (error: object, variables: CreateProjectWorkflowRequest) => void;
-};
-
-export const useCreateProjectWorkflowRequestMutation = (
-    mutationProps?: CreateProjectWorkflowRequestMutationProps
-) => {
-    return useMutation({
-        mutationFn: (request: CreateProjectWorkflowRequest) => {
-            return new ProjectApi().createProjectWorkflow(request);
-        },
-        onError: mutationProps?.onError,
-        onSuccess: mutationProps?.onSuccess,
-    });
-};
-
 type DeleteProjectMutationProps = {
     onSuccess?: (result: void, variables: number) => void;
     onError?: (error: object, variables: number) => void;
@@ -62,22 +34,6 @@ export const useDeleteProjectMutation = (
     useMutation({
         mutationFn: (id: number) => {
             return new ProjectApi().deleteProject({id: id});
-        },
-        onError: mutationProps?.onError,
-        onSuccess: mutationProps?.onSuccess,
-    });
-
-type DeleteWorkflowMutationProps = {
-    onSuccess?: (result: void, variables: DeleteProjectWorkflowRequest) => void;
-    onError?: (error: object, variables: DeleteProjectWorkflowRequest) => void;
-};
-
-export const useDeleteWorkflowMutation = (
-    mutationProps?: DeleteWorkflowMutationProps
-) =>
-    useMutation({
-        mutationFn: (request: DeleteProjectWorkflowRequest) => {
-            return new WorkflowApi().deleteProjectWorkflow(request);
         },
         onError: mutationProps?.onError,
         onSuccess: mutationProps?.onSuccess,
@@ -96,25 +52,6 @@ export const useDuplicateProjectMutation = (
             return new ProjectApi().duplicateProject({
                 id: id,
             });
-        },
-        onError: mutationProps?.onError,
-        onSuccess: mutationProps?.onSuccess,
-    });
-
-type DuplicateWorkflowMutationProps = {
-    onSuccess?: (
-        result: WorkflowModel,
-        variables: DuplicateWorkflowRequest
-    ) => void;
-    onError?: (error: object, variables: DuplicateWorkflowRequest) => void;
-};
-
-export const useDuplicateWorkflowMutation = (
-    mutationProps?: DuplicateWorkflowMutationProps
-) =>
-    useMutation({
-        mutationFn: (request: DuplicateWorkflowRequest) => {
-            return new WorkflowApi().duplicateWorkflow(request);
         },
         onError: mutationProps?.onError,
         onSuccess: mutationProps?.onSuccess,
@@ -153,41 +90,6 @@ export const useUpdateProjectMutation = (
                 id: projectModel.id!,
                 projectModel,
             });
-        },
-        onError: mutationProps?.onError,
-        onSuccess: mutationProps?.onSuccess,
-    });
-
-type UpdateProjectTagsMutationProps = {
-    onSuccess?: (result: void, variables: UpdateProjectTagsRequest) => void;
-    onError?: (error: object, variables: UpdateProjectTagsRequest) => void;
-};
-
-export const useUpdateProjectTagsMutation = (
-    mutationProps?: UpdateProjectTagsMutationProps
-) =>
-    useMutation({
-        mutationFn: (request: UpdateProjectTagsRequest) => {
-            return new ProjectTagApi().updateProjectTags(request);
-        },
-        onError: mutationProps?.onError,
-        onSuccess: mutationProps?.onSuccess,
-    });
-
-type UpdateWorkflowMutationProps = {
-    onSuccess?: (
-        result: WorkflowModel,
-        variables: UpdateWorkflowRequest
-    ) => void;
-    onError?: (error: object, variables: UpdateWorkflowRequest) => void;
-};
-
-export const useUpdateWorkflowMutation = (
-    mutationProps?: UpdateWorkflowMutationProps
-) =>
-    useMutation({
-        mutationFn: (request: UpdateWorkflowRequest) => {
-            return new WorkflowApi().updateWorkflow(request);
         },
         onError: mutationProps?.onError,
         onSuccess: mutationProps?.onSuccess,
