+++ client/src/queries/connections.queries.ts
@@ -1,3 +1,4 @@
+/* eslint-disable sort-keys */
 import {
     ConnectionApi,
     ConnectionModel,
@@ -5,11 +6,6 @@ import {
     GetConnectionsRequest,
     TagModel,
 } from '@/middleware/helios/connection';
-import {
-    GetOAuth2AuthorizationParametersRequestModel,
-    OAuth2AuthorizationParametersModel,
-    Oauth2Api,
-} from '@/middleware/hermes/configuration';
 import {useQuery} from '@tanstack/react-query';
 
 export const ConnectionKeys = {
@@ -18,9 +14,6 @@ export const ConnectionKeys = {
         ...ConnectionKeys.connections,
         filters,
     ],
-    connectionOAuth2AuthorizationParameters: (
-        request: GetOAuth2AuthorizationParametersRequestModel
-    ) => [...ConnectionKeys.connections, request],
     connectionTags: ['connectionTags'] as const,
     connections: ['connections'] as const,
 };
@@ -29,28 +22,14 @@ export const useGetConnectionsQuery = (
     filters: GetConnectionsRequest,
     enabledCondition?: boolean
 ) =>
-    useQuery<ConnectionModel[], Error>(
-        ConnectionKeys.connectionList(filters),
-        () => new ConnectionApi().getConnections(filters),
-        {enabled: false || enabledCondition}
-    );
+    useQuery<ConnectionModel[], Error>({
+        queryKey: ConnectionKeys.connectionList(filters),
+        queryFn: () => new ConnectionApi().getConnections(filters),
+        enabled: false || enabledCondition,
+    });
 
 export const useGetConnectionTagsQuery = () =>
-    useQuery<TagModel[], Error>(ConnectionKeys.connectionTags, () =>
-        new ConnectionTagApi().getConnectionTags()
-    );
-
-export const useGetOAuth2AuthorizationParametersQuery = (
-    request: GetOAuth2AuthorizationParametersRequestModel,
-    enabled: boolean
-) =>
-    useQuery<OAuth2AuthorizationParametersModel, Error>(
-        ConnectionKeys.connectionOAuth2AuthorizationParameters(request),
-        () =>
-            new Oauth2Api().getOAuth2AuthorizationParameters({
-                getOAuth2AuthorizationParametersRequestModel: request,
-            }),
-        {
-            enabled,
-        }
-    );
+    useQuery<TagModel[], Error>({
+        queryKey: ConnectionKeys.connectionTags,
+        queryFn: () => new ConnectionTagApi().getConnectionTags(),
+    });
