+++ client/src/queries/projectInstances.queries.ts
@@ -1,8 +1,7 @@
+/* eslint-disable sort-keys */
 import {
     ProjectInstanceApi,
     ProjectInstanceModel,
-    ProjectInstanceTagApi,
-    TagModel,
 } from '@/middleware/helios/configuration';
 import {useQuery} from '@tanstack/react-query';
 
@@ -11,7 +10,6 @@ export const ProjectInstanceKeys = {
         projectId?: number;
         tagId?: number;
     }) => [...ProjectInstanceKeys.projectInstances, filters],
-    projectInstanceTags: ['projectInstanceTags'] as const,
     projectInstances: ['projectInstances'] as const,
 };
 
@@ -19,12 +17,7 @@ export const useGetProjectInstancesQuery = (filters: {
     projectId?: number;
     tagId?: number;
 }) =>
-    useQuery<ProjectInstanceModel[], Error>(
-        ProjectInstanceKeys.filteredProjectInstances(filters),
-        () => new ProjectInstanceApi().getProjectInstances(filters)
-    );
-
-export const useGetProjectInstanceTagsQuery = () =>
-    useQuery<TagModel[], Error>(ProjectInstanceKeys.projectInstanceTags, () =>
-        new ProjectInstanceTagApi().getProjectInstanceTags()
-    );
+    useQuery<ProjectInstanceModel[], Error>({
+        queryKey: ProjectInstanceKeys.filteredProjectInstances(filters),
+        queryFn: () => new ProjectInstanceApi().getProjectInstances(filters),
+    });
