+++ client/src/queries/projectTags.quries.ts
@@ -1,0 +1,13 @@
+/* eslint-disable sort-keys */
+import {ProjectTagApi, TagModel} from '@/middleware/helios/configuration';
+import {useQuery} from '@tanstack/react-query';
+
+export const ProjectTagKeys = {
+    projectTags: ['projectTags'] as const,
+};
+
+export const useGetProjectTagsQuery = () =>
+    useQuery<TagModel[], Error>({
+        queryKey: ProjectTagKeys.projectTags,
+        queryFn: () => new ProjectTagApi().getProjectTags(),
+    });
